Server/requirements.txt
GoogleBard
EdgeGPT
requests
